src/mirza/streamlit/requirements.txt
streamlit==1.37.1
psycopg2-binary==2.9.9
pandas==2.2.2
SQLAlchemy==2.0.32
plotly==5.23.0
python-dotenv==1.0.1
